+++ PriemLib/Statistics/RegionAbitStatstics_Reversed.cs
@@ -153,9 +153,9 @@ WHERE extPerson.RegionId=@RegionId AND extAbit.StudyLevelGroupId=@StudyLevelGrou
                              {
                                  FacultyName = rw.Field<string>("FacultyName"),
                                  StudyBasisId = rw.Field<int>("StudyBasisId"),
-                                 MinSUM = rw.Field<int?>("MinSUM"),
+                                 MinSUM = rw.Field<decimal?>("MinSUM"),
                                  AvgSUM = rw.Field<double?>("AvgSUM") ?? 0d,
-                                 MaxSUM = rw.Field<int?>("MaxSUM")
+                                 MaxSUM = rw.Field<decimal?>("MaxSUM")
                              }).ToList();
 
             DataTable src_tbl = new DataTable();
@@ -170,13 +170,12 @@ WHERE extPerson.RegionId=@RegionId AND extAbit.StudyLevelGroupId=@StudyLevelGrou
             src_tbl.Columns.Add("% абит. г/б", typeof(float));
             src_tbl.Columns.Add("% абит. дог", typeof(float));
 
-            src_tbl.Columns.Add("Мин. балл г/б", typeof(int));
-            src_tbl.Columns.Add("Мин. балл дог", typeof(int));
+            src_tbl.Columns.Add("Мин. балл г/б", typeof(decimal));
+            src_tbl.Columns.Add("Мин. балл дог", typeof(decimal));
             src_tbl.Columns.Add("Сред. балл г/б", typeof(double));
             src_tbl.Columns.Add("Сред. балл дог", typeof(double));
-            src_tbl.Columns.Add("Макс. балл г/б", typeof(int));
-            src_tbl.Columns.Add("Макс. балл дог", typeof(int));
-
+            src_tbl.Columns.Add("Макс. балл г/б", typeof(decimal));
+            src_tbl.Columns.Add("Макс. балл дог", typeof(decimal));
 
             int iCntAbitBudzh = data_counts.Where(x => x.StudyBasisId == 1).Select(x => x.CNT_abit).Sum();
             int iCntAbitPlatn = data_counts.Where(x => x.StudyBasisId == 2).Select(x => x.CNT_abit).Sum();
